refactor: extract plane data types from pipeline.rs into planes.rs
YuvPlanes, Nv12Planes, BgraPlanes, StridedYuvPlanes, FramePlaneView,
and copy_plane_tight are pure data types with no GPU dependency. They
were buried in pipeline.rs (the GPU orchestrator) despite being
imported by many modules across the crate.

Now in their own module with re-exports from pipeline.rs for
backwards compatibility within the crate.

pipeline.rs: 1,102 -> 920 lines (-182).
