+++ crates/reco-core/src/pipeline.rs
@@ -31,192 +31,7 @@ use crate::viewport::{ResolvedViewport, ViewportConfig};
 
 use thiserror::Error;
 
-/// Borrowed YUV420P plane references for pipeline input.
-///
-/// Tightly packed (no stride padding):
-/// - `y`: `width × height` bytes
-/// - `u`: `(width/2) × (height/2)` bytes
-/// - `v`: `(width/2) × (height/2)` bytes
-pub struct YuvPlanes<'a> {
-    /// Y (luma) plane, full resolution.
-    pub y: &'a [u8],
-    /// U (Cb) plane, half resolution.
-    pub u: &'a [u8],
-    /// V (Cr) plane, half resolution.
-    pub v: &'a [u8],
-}
-
-/// A single stride-aware plane view.
-///
-/// External frameworks (OBS `obs_source_frame`, V4L2, GStreamer buffers)
-/// typically expose planes as a pointer plus a row stride that may exceed
-/// the plane width (padding). [`YuvPlanes`] assumes tight packing, so a
-/// consumer that receives padded data has to repack before calling
-/// [`StitchPipeline::render_to_target`]. [`StridedYuvPlanes::copy_into`]
-/// does the repack into a caller-owned buffer without reallocating on
-/// every frame.
-pub struct FramePlaneView<'a> {
-    /// Plane pixel bytes. Must have length at least `stride * height`.
-    pub data: &'a [u8],
-    /// Bytes per row, including any trailing padding.
-    pub stride: u32,
-    /// Plane width in pixels (bytes per row of usable data).
-    pub width: u32,
-    /// Plane height in rows.
-    pub height: u32,
-}
-
-/// Stride-aware YUV420P planes, suitable for hardware-decoded frames
-/// or framework callbacks that expose raw pointers + strides.
-///
-/// Use [`Self::copy_into`] to produce a tightly packed [`YuvPlanes`]
-/// view ready for [`StitchPipeline::render_to_target`].
-pub struct StridedYuvPlanes<'a> {
-    /// Y (luma) plane, full resolution.
-    pub y: FramePlaneView<'a>,
-    /// U (Cb) plane, half resolution per dimension.
-    pub u: FramePlaneView<'a>,
-    /// V (Cr) plane, half resolution per dimension.
-    pub v: FramePlaneView<'a>,
-}
-
-impl StridedYuvPlanes<'_> {
-    /// Repack the strided planes into a caller-owned tight buffer and
-    /// return a [`YuvPlanes`] view into it.
-    ///
-    /// The `buffer` is resized to `y_len + u_len + v_len` bytes. Cache
-    /// the buffer across frames to avoid per-frame allocation.
-    pub fn copy_into<'b>(&self, buffer: &'b mut Vec<u8>) -> YuvPlanes<'b> {
-        let y_len = (self.y.width as usize) * (self.y.height as usize);
-        let u_len = (self.u.width as usize) * (self.u.height as usize);
-        let v_len = (self.v.width as usize) * (self.v.height as usize);
-        buffer.resize(y_len + u_len + v_len, 0);
-        {
-            let (y_dst, rest) = buffer.split_at_mut(y_len);
-            let (u_dst, v_dst) = rest.split_at_mut(u_len);
-            copy_plane_tight(&self.y, y_dst);
-            copy_plane_tight(&self.u, u_dst);
-            copy_plane_tight(&self.v, v_dst);
-        }
-        let (y, rest) = buffer.split_at(y_len);
-        let (u, v) = rest.split_at(u_len);
-        YuvPlanes { y, u, v }
-    }
-}
-
-/// Copy `src` rows into `dst`, skipping stride padding.
-///
-/// `dst` must be exactly `src.width * src.height` bytes long.
-/// Rows where `stride == width` are copied in a single `copy_from_slice`.
-///
-/// B-24 (deep-review-2026-04-18): the loop below slices
-/// `src.data[row*stride..row*stride + width]`, which panics if
-/// `stride < width` (overlaps next row) or if the source data is
-/// shorter than `stride*height` (out of bounds). Previously these
-/// preconditions were only `debug_assert!`, so release builds
-/// produced a panic deep in the render path instead of a safe
-/// fallback. Now the preconditions are runtime-checked and a
-/// malformed plane falls back to zero-filling the destination plus a
-/// warn log so the pipeline continues with a black plane rather than
-/// aborting.
-fn copy_plane_tight(src: &FramePlaneView<'_>, dst: &mut [u8]) {
-    let width = src.width as usize;
-    let height = src.height as usize;
-    let stride = src.stride as usize;
-
-    if dst.len() != width.saturating_mul(height) {
-        log::warn!(
-            "copy_plane_tight: dst {} bytes != width*height {} bytes; zero-filling",
-            dst.len(),
-            width.saturating_mul(height),
-        );
-        dst.fill(0);
-        return;
-    }
-    if stride < width {
-        log::warn!(
-            "copy_plane_tight: stride {stride} < width {width}; malformed FramePlaneView, \
-             zero-filling plane",
-        );
-        dst.fill(0);
-        return;
-    }
-    if src.data.len() < stride.saturating_mul(height) {
-        log::warn!(
-            "copy_plane_tight: source buffer {} bytes < stride*height {} bytes; \
-             zero-filling plane",
-            src.data.len(),
-            stride.saturating_mul(height),
-        );
-        dst.fill(0);
-        return;
-    }
-
-    if stride == width {
-        // Fast path: source is already tight; one memcpy.
-        dst.copy_from_slice(&src.data[..width * height]);
-        return;
-    }
-
-    for row in 0..height {
-        let src_start = row * stride;
-        let dst_start = row * width;
-        dst[dst_start..dst_start + width].copy_from_slice(&src.data[src_start..src_start + width]);
-    }
-}
-
-/// Borrowed NV12 plane references for pipeline input.
-///
-/// Tightly packed (no stride padding):
-/// - `y`: `width × height` bytes
-/// - `uv`: `width × (height/2)` bytes (interleaved U,V)
-pub struct Nv12Planes<'a> {
-    /// Y (luma) plane, full resolution.
-    pub y: &'a [u8],
-    /// Interleaved UV (CbCr) plane, half resolution in each dimension.
-    pub uv: &'a [u8],
-}
-
-/// Borrowed packed BGRA/RGBA plane for pipeline input.
-///
-/// Tightly packed, 4 bytes per pixel in (R, G, B, A) byte order - the
-/// shader samples `rgba.rgb` directly, so callers with BGRA data need
-/// to swizzle bytes at upload time (see
-/// [`BgraPlanes::from_bgra_swizzle_into`]). Used for OBS Browser
-/// Source, screen capture, WebRTC ingest, and any other consumer whose
-/// source frames are already sRGB-domain.
-pub struct BgraPlanes<'a> {
-    /// Packed RGBA bytes, length `width * height * 4`.
-    pub rgba: &'a [u8],
-}
-
-impl<'a> BgraPlanes<'a> {
-    /// Wrap an already-RGBA byte slice without copying.
-    ///
-    /// Length must be exactly `width * height * 4`; callers are expected
-    /// to match the dimensions the pipeline was constructed with.
-    pub fn from_rgba(rgba: &'a [u8]) -> Self {
-        Self { rgba }
-    }
-
-    /// Swizzle a BGRA-ordered source into a caller-owned `Vec<u8>` and
-    /// return a [`BgraPlanes`] borrowing from it.
-    ///
-    /// OBS and most desktop compositors deliver BGRA; the shader expects
-    /// RGBA, so this helper performs the cheap byte-reorder once per
-    /// frame. Cache the buffer across frames to avoid per-frame
-    /// allocation (the helper `resize`s in place).
-    pub fn from_bgra_swizzle_into(bgra: &[u8], buffer: &'a mut Vec<u8>) -> Self {
-        buffer.resize(bgra.len(), 0);
-        for (src, dst) in bgra.chunks_exact(4).zip(buffer.chunks_exact_mut(4)) {
-            dst[0] = src[2]; // R <- B
-            dst[1] = src[1]; // G
-            dst[2] = src[0]; // B <- R
-            dst[3] = src[3]; // A
-        }
-        Self { rgba: buffer }
-    }
-}
+pub use crate::planes::{BgraPlanes, FramePlaneView, Nv12Planes, StridedYuvPlanes, YuvPlanes};
 
 /// Errors from the stitch pipeline. `Clone + Send + Sync` so consumers
 /// posting results to worker threads can carry the typed error.
@@ -916,6 +731,7 @@ impl StitchPipeline {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::planes::copy_plane_tight;
 
     /// Build a test plane where row `r` contains byte value `r` for the first
     /// `width` bytes, followed by `0xFF` padding up to `stride`.
